+++ geeadd/batch_mover.py
@@ -18,12 +18,19 @@ __copyright__ = """
 
 """
 __license__ = "Apache 2.0"
-import ee
 import os
 
+import ee
+
+
+def camel_case(s):
+    words = s.split()
+    temp = words
+    res = ' '.join(ele.title() for ele in temp)
+    return res
 
 # Image copy
-def image_move(initial, replace_string, replaced_string, fpath):
+def image_move(initial, replace_string, replaced_string, fpath,ftype):
     ee.Initialize()
     if replace_string == replaced_string or replace_string == None:
         final = fpath
@@ -41,7 +48,7 @@ def image_move(initial, replace_string, replaced_string, fpath):
 
 
 # Table copy
-def table_move(initial, replace_string, replaced_string, fpath):
+def table_move(initial, replace_string, replaced_string, fpath,ftype):
     ee.Initialize()
     if replace_string == replaced_string or replace_string == None:
         final = fpath
@@ -49,9 +56,9 @@ def table_move(initial, replace_string, replaced_string, fpath):
         final = initial.replace(replace_string, replaced_string)
     try:
         if ee.data.getAsset(final):
-            print("Table already moved: {}".format(final))
-    except Exception:
-        print("Moving table to {}".format(final))
+            print(f"{camel_case(ftype)} already moved: {final}")
+    except Exception as e:
+        print(f"Moving {camel_case(ftype)} to {final}")
         try:
             ee.data.renameAsset(initial, final)
         except Exception as e:
@@ -170,8 +177,14 @@ def mover(path, fpath):
                                 child["name"], replace_string, replaced_string, fpath
                             )
                         elif child["type"].lower() == "table":
+                            ftype = "table"
                             table_move(
                                 child["name"], replace_string, replaced_string, fpath
+                            )
+                        elif child["type"].lower() == "feature_view":
+                            ftype = "feature view"
+                            table_move(
+                                child["name"], replace_string, replaced_string, fpath,ftype
                             )
                     print("")
             elif ee.data.getAsset(path)["type"].lower() == "image":
@@ -235,12 +248,21 @@ def mover(path, fpath):
                     except Exception as e:
                         print(e)
             elif ee.data.getAsset(path)["type"].lower() == "table":
+                ftype = "table"
                 path = ee.data.getAsset(path)["name"]
                 replace_string = None
                 replaced_string = ee.data.getAsset(
                     "/".join(fpath.split("/")[:-1]) + "/"
                 )["name"]
-                table_move(path, replace_string, replaced_string, fpath)
+                table_move(path, replace_string, replaced_string, fpath,ftype)
+            elif ee.data.getAsset(path)["type"].lower() == "feature_view":
+                ftype = "feature view"
+                path = ee.data.getAsset(path)["name"]
+                replace_string = None
+                replaced_string = ee.data.getAsset(
+                    "/".join(fpath.split("/")[:-1]) + "/"
+                )["name"]
+                table_move(path, replace_string, replaced_string, fpath,ftype)
     except Exception as e:
         print(e)
         print("Initial path {} not found".format(path))